v2
Had to create a method to refund all trades entered against an item listed since multiple trades could be entered. When an item sell, if all of the trades aren't refunded, then the remaining trades will stay on contract for that tradeId which will be recycled. 
Added a method to grab all of the trades by itemId, then iterate through them to refund them all, lines 502-505.

I also had to add an internal _refundTradeFromSale function in order to prevent an access control error.