new random token Id func
created new random Id following the offline discussion to use an array of tokenIds and pop off the last one.
Added a function to push additional tokenIds onto the array, which updates the totalNfts also. 
The totalNfts are used as the nonce to create the random tokenId, and when the last item is shifted into the chosen tokens place and popped off, the totalnfts is updated to the new length.
Removed the check for a zero random number as it is a valid choice in the array.