adjusted the tokenId update function
In order to avoid potential human err, I updated the setTokenId function to be incremented with a totalNfts uint count. added a variable to keep track of the remaining amount of NFTs and totalNfts. When we set a new batch of tokenIds we just enter the amount created, and it will increment the totalNfts count and push that onto the array for use.