Added toWhom[] array of addresses for minting.
In order to send the honorary nfts on first deployment while keeping track of the ids and addresses, I added the array of addresses to mint to. It could be useful in the future also to mint and send to others.
I also added fetchNFTsByAddress and an internal _addressToUserCount[] mapping to keep track of how many a user minted and retrieve that memory amount of items for the fetch.
Tests passed and I have all of the addresses that gifted to help out in the testing array.